Update d2l_schedule_training.html
Added necessary  {% csrf_token %} and method = "POST", and updated ID and Names, and other attributes of the input for the form. This form properly function in as much that it will send to the clients the information from the forms. settings.py will need to include client info for receiving form submissions